experiments/temporal-models/pyro-detector-baseline/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pyro_detector_baseline"]

[project]
name = "pyro-detector-baseline"
version = "0.1.0"
description = "Production pyro-predictor wrapped as TemporalModel baseline for comparison"
requires-python = ">=3.11"
dependencies = [
    "pyrocore",
    "pyro-predictor",
    "huggingface-hub>=0.24",
    "matplotlib>=3.8",
    "numpy>=1.26,<2",
    "opencv-python-headless>=4.8",
    "onnxruntime>=1.17",
    "pandas>=2.1",
    "pillow>=10.0",
    "pyyaml>=6.0",
    "seaborn>=0.13",
    "tqdm>=4.66",
    "ncnn>=1.0.20260114",
]

[tool.uv.sources]
pyrocore = { path = "../../../lib/pyrocore" }
pyro-predictor = { git = "https://github.com/pyronear/pyro-engine", subdirectory = "pyro-predictor" }

[dependency-groups]
dev = [
    "dvc[s3]>=3.56",
    "nbqa>=1.9",
    "nbstripout>=0.8",
    "pytest>=8.0",
    "ruff>=0.9",
]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM", "PLC0415"]

[tool.ruff.format]
quote-style = "double"
